Create SECURITY.md
Hello Team,

As the continued popularity of this project increases we should be taking our projects goals in mind with concerns to privacy and security. In the near future I would like to at least start testing the codebase with SAST(semgrep) and Synk.io (Software Composition Analysys) To highlight any potential weaknesses in the security of the product. 

I am happy to start contributing from the security side.