Move svelte projet files in a UUID base directory
A random but for now static UUID is used to relocate the frontend project.
We first move all sveltes source files in that directory, and add the necessary symbolic links to avoid changing all imports.
The project at that stage is broken.
Further commits will files all files to reflect the change.
A deployment stage will be able to replace the UUID with any PATH.