Add run-pip.bat script to install Python 3.11 and run open-webui
This commit adds a new run-pip.bat script that automates the installation of Python 3.11 and the execution of the open-webui application. The script checks if Python 3.11 is already installed and installs it if necessary. It then creates and activates a virtual environment named "myenv". After that, it installs the open-webui package using pip and runs the open-webui application.