Add support for cookie based sticky sessions
When multiple instances of vLLM/OLLAMA are deployed behind a loadbalancer, sticky sessions are enabled in the load balancer to route requests from a client to the backend it originally hit to ensure KV-Cache reuse. One way to implement sticky sessions is via a cookie that the loadbalancer sends to the client.

This adds support for setting an env var named PASSTHROUGH_SESSION_COOKIES that stores a comma separated list of cookie names. Open-Webui already sends back any cookies it receives from the LM backend to the browser. If this var is set, any cookies matching this names in this list that it receives from the browser in subsequent requests will be forwarded on to the LM backend