feat: A new parameter model called "native_tool_call" is present. (squashed commit for easier review)
If false (the default), the behaviour when there are tools enabled
 is the same: a prompt is built which asks the model to generate
a function call compatible with the presented tools. The output is
the added to the context.

If true, the native mechanism that the API has is used instead. That
means that the responses have to be inspected to detect if a tool_call
has to be handled and act accordingly. This allows more complex
 behaviours where the model can try several strategies and call
tools knowing the output of others.

Ollama tool calling API only supports non streaming requests.
In case both streaming and native_tool call are present, native
tool calling is silently ignored.