Update requirements.txt to support newer OpenAI turbo model
**Is your feature request related to a problem? Please describe.**
The current version of the faster-whisper used by Open WebUI does not support the newer OpenAI turbo model.

**Describe the solution you'd like**
Update the faster-whisper dependency within Open WebUI to version 1.1.1, which supports the turbo model. This would allow users to select and utilize this newer model for transcriptions within the Open WebUI interface.

**Additional context**
The turbo model is expected to offer 8x speed improvements compared to the large model. I have updated the faster-whisper to version 1.1.1 and confirmed that it works flawlessly with the turbo model within Open WebUI.