chore(app): refactor Sidebar from +layout to +pages
Prepare refactoring Sidebar new chat button to dispatch event which can
be listened to to call new chat.

== Background

To start a new chat the Sidebar component uses getElementById() to locate
a new chat button within the Navbar (yes) to call click().

This is bad in two ways:

1. It creates depedencies across the application (i.e. the Navbar new
   chat button can not be removed without breaking the Sidebar new chat
   button)

2. It makes the functionality harder to read/trace.

== Solution

Since it's not possible to propagate events into a slot it's not
possible to bubble it up and then down into the slot and Chat.

Therefor Chat moved into the children pages.