chore(Chat): start new chat though explicit function call
... not through globally queried new chat button, which is clicked.

== Background

To start a new chat the Sidebar component uses getElementById() to locate
a new chat button within the Navbar (yes) to call click().

This is bad in two ways:

1. It creates depedencies across the application (i.e. the Navbar new
   chat button can not be removed without breaking the Sidebar new chat
   button)

2. It makes the functionality harder to read/trace.

== Solution

* A newChat() function is exported on Chat
* Sidebar now dispatches a "new-chat" event
* Consuming pages listen to the event an call newChat()

== Note

One would think that after navigation to / via goto() it's redundant to
explicitly call initNewChat(), but a second click on the new chat
buttons would not work. (Unclear why)