chore: improve config.json migration
Checking if the file exists and then attempting to open it could cause the program to crash if the file is moved or deleted between those operations. While this is extremely unlikely to occur naturally we should fix it anyway.

load_json_config was only used by the migration and doesn't need its own function.

We previously saved the loaded config in the `data` variable. This has a global scope and would not get removed by the gc. We can fix this by moving the migration into its own function.