c/feat(tailwind.config): add xs breakpoint
To have a dedicated breakpoint to expand the explore page agents.

I also attempted to follow the Documented [1] theming guide and prefer
the in-CSS theming, yet it does not work adding breakpoints like in [1]
to the app.css (even with build restart).

Breakpoint value was chosen right before the second card break to the
next line.

[1]: https://tailwindcss.com/docs/responsive-design#customizing-your-theme

Refs: PRODAI-105