c/feat(explore): integrate login/register overlay
The overlay is presented when the user has no session within Open WebUI.

If the user choses to signup the agent or prompt selection is discarded
and the user is sent to the signup URL as the user would most likely not
remember what they picked.

If the user choses to login, they're sent down the usual auth process
(/auth), which then starts OIDC.

If the user comes back through the OIDC login process it's checked
whether they went through the registration. In that case the user is
sent back to the expore page.

Refs: PRODAI-148