FIX: num_keep param explanation/exemplified (en-US)
Def. n_keep:
It's used when the n_ctx limit is reached. A new prompt will be constructed with the first n_keep characters of the original prompt plus the second half of the output to free up space for more conversation.
It's the amount of tokens after your system prompt that are always kept, when the start of the history is thrown away.

First N, not last N

In Params Configs it's wrong exemplified, as:
"This option controls how many tokens are preserved when refreshing the context. For example, if set to 2, the last 2 tokens of the conversation context will be retained. Preserving context can help maintain the continuity of a conversation, but it may reduce the ability to respond to new topics."

When have to be, as:
"This option controls how many tokens are preserved when refreshing the context. For example, if set to 2, the first 2 tokens of the conversation context will be retained. Preserving context can help maintain the continuity of a conversation, but it may reduce the ability to respond to new topics."