c/fix(auths): change logout from end_session_endpoint to ionos_logout_url
== Background

With OIDC enabled, the user was sent to the end_session_endpoint
provided by the OIDC discovery. This, however, is not supported by the
IONOS auth server.

With the previuous implementation the user was sent to a wrong URL which
did not invalidate the session at the Auth server, but showed a Basic
Auth dialog for the unknown URL.

== Fix

A custom configuration is introduced where users shall be redirected to.

The target location accepts a "redirect_url" parameter to send the user
to after successful logout. We send the user back to the Startpage
(explore).

If the configuration is not present we'll fall back to
end_session_endpoint (mostly for testing in Minikube).