Merged PR 176: Accessibility: Implemented Modal rework to allow focus-trap functionality to be refactored.
implemented Modal rework to allow focus-trap functionality to be refactored.
The following Modals do not have proper focus trap return functionality, as it does not work for components that are in dropdowns as the focus element disappears.
-SettingsModal.svelte
-ShareChatModal.svelte
-ShortcutsModal.svelte
-ArchivedChatsModal.svelte

The delete Folder does not have focus trap set up.
![image.png](https://dev.azure.com/ESDC-AICoE/8d2e9668-3fce-49b9-98d1-cdde6d0ae324/_apis/git/repositories/b039c9ad-6cbd-4479-a2e6-c848fb54ec92/pullRequests/176/attachments/image.png)
The delete Chat does not have focus trap set up
![image (2).png](https://dev.azure.com/ESDC-AICoE/8d2e9668-3fce-49b9-98d1-cdde6d0ae324/_apis/git/repositories/b039c9ad-6cbd-4479-a2e6-c848fb54ec92/pullRequests/176/attachments/image%20%282%29.png)

Related work items: #1738, #1836