c/feat(main): custom middleware to prevent caching for some paths
== Problem

Currently no caching rules are defined as headers leaving the caching up
to the browser vendor.

Unless a user makes a hard reload of the page, they may not load the
newest version of the application.

== Solution

To ensure that the entrypoints are always the newest version, caching is
explicitly prevented using Expires and Cache-Control headers.

== Paths

The configured paths are the most common places users end up and
potentially revisit the SPA from.

Refs: PRODAI-51