c/fix(github): check for job to terminate after following logs
Sometimes the Job was successful but the following status check failed.
By analyzing the job status debug output it turned out that the
container status was still running.

Assumption: the container needs a couple of seconds after log following
ended to turn to status "terminated".

We now wait for it no longer be in running status.