feat: improve "expire on views" UX in share modal
Previously, the input field for the number of views would default to 1, and there was no client-side validation to prevent you from entering a number less than or equal to the current view count. This could lead to confusion when the link would be immediately invalid upon creation.

This change introduces two improvements:
1.  When you select the "Expire after a number of views" option, the input field now automatically defaults to a valid number (current view count + 1) if the existing value is insufficient.
2.  If you manually enter an invalid number (less than or equal to the current view count) and click away from the input field, the value is automatically corrected to a valid number, and a toast notification is displayed to inform you of the change.