FIX: Show Content of Uploaded File in FileItemModal.svelte
FIX: Show Content of Uploaded File in FileItemModal.svelte

Problem:
Content of uploaded files added to the message (e.g. text or xlsx) is not displayed in the FileItemModal, it simply appears as  "No Content". However, the embedded is successful and the content can be displayed simply by clicking the file link in the modal.

This feature was working propertly in v0.6.22 with the same FileItem & FileItemModal files, but not in the dev v.0.6.26

With this update, that repopulate item.file.data.content, the content is displayed correctly.

Note: This is a quick fix to fit the current structure, but I think try/catch clauses are needed here.