@FIR-956: The model download from huggingface and ollama was not showing (#74)
right status
This change handles the hugging face and ollama download. The root cause
of the problem was extra yield and unnencessary redirection and await
while stream handling when uploading model to FPGA through flask.

The test results are as following, now when the upload to FPGA is
finished and down load is complete there is no Cancelled or other status
displayed which was confusion earlier.

Co-authored-by: Ashish Trivedi <atrivedi@fpga3.tsavoritesi.net>