refac: add enabled/disabled tooltips to all toggle switches
This commit adds "Enabled" and "Disabled" tooltips to all instances of the `Switch.svelte` component throughout the application.

This was achieved by adding the `tooltip={true}` prop to the `Switch` component. In many cases, this involved refactoring the code to remove a redundant `<Tooltip>` wrapper component in favor of the more concise and standardized prop-based approach.

This change improves user experience by providing clear feedback on the state of toggle switches.